requirements.txt
boto3
pandas
numpy
requests
apache-airflow
praw
apache-airflow-providers-amazon


